REF: move infer function to a __main__ module
