+++ ptychonn/__init__.py
@@ -6,4 +6,4 @@ except PackageNotFoundError:
     # package is not installed
     pass
 
-from ptychonn._infer.workflow import infer
+from ptychonn._infer.__main__ import infer
